# internal/controlplane/server/http_agents.go
package server

import (
	"errors"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
	"github.com/lost-coder/panvex/internal/controlplane/auth"
	"github.com/lost-coder/panvex/internal/controlplane/storage"
)

// msgMissingAgentID is returned when a handler is called without the agent
// id path segment; hoisted to avoid duplicating the literal (go:S1192).
const msgMissingAgentID = "missing agent id"

type renameAgentRequest struct {
	NodeName string `json:"node_name"`
}

type updateAgentFleetGroupRequest struct {
	FleetGroupID string `json:"fleet_group_id"`
}

func (s *Server) handleRenameAgent() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		session, user, err := s.requireSession(r)
		if err != nil {
			writeErrorLogged(r.Context(), w, http.StatusUnauthorized, "unauthorized", err)
			return
		}

		agentID, nodeName, ok := decodeRenameAgentRequest(w, r)
		if !ok {
			return
		}

		if !s.checkRenameAgentScope(w, r, user, agentID) {
			return
		}

		if !s.persistAgentNodeName(w, r, agentID, nodeName) {
			return
		}

		agent, oldName, ok := s.applyAgentRename(w, agentID, nodeName)
		if !ok {
			return
		}

		s.appendAuditWithContext(r.Context(), session.UserID, "agents.rename", agentID, map[string]any{
			"old_name": oldName,
			"new_name": nodeName,
		})

		writeJSON(w, http.StatusOK, agent)
	}
}

// decodeRenameAgentRequest extracts and validates the agent id + new
// node name. Writes a 400 / response on failure and returns ok=false.
func decodeRenameAgentRequest(w http.ResponseWriter, r *http.Request) (string, string, bool) {
	agentID := chi.URLParam(r, "id")
	if agentID == "" {
		writeError(w, http.StatusBadRequest, msgMissingAgentID)
		return "", "", false
	}

	var req renameAgentRequest
	if err := decodeJSON(r, &req); err != nil {
		writeErrorLogged(r.Context(), w, http.StatusBadRequest, "invalid request body", err)
		return "", "", false
	}

	nodeName := strings.TrimSpace(req.NodeName)
	if nodeName == "" {
		writeError(w, http.StatusBadRequest, "node_name is required")
		return "", "", false
	}
	return agentID, nodeName, true
}

// checkRenameAgentScope confirms the target agent exists in memory and
// the caller's fleet scope covers it. R-S-14: scope-check before any
// write so an out-of-scope rename leaks no information.
func (s *Server) checkRenameAgentScope(w http.ResponseWriter, r *http.Request, user auth.User, agentID string) bool {
	// Verify the agent exists in memory before touching the store so a
	// 404 does not leave an orphaned store update.
	existing, exists := s.live.Get(agentID)
	if !exists {
		writeError(w, http.StatusNotFound, msgAgentNotFound)
		return false
	}

	scope, ok := s.requireFleetScope(w, r, user)
	if !ok {
		return false
	}
	if !scope.IsAllowed(existing.FleetGroupID) {
		writeError(w, http.StatusNotFound, msgAgentNotFound)
		return false
	}
	return true
}

// persistAgentNodeName writes the new node name to storage when one is
// configured, returning false (after writing the HTTP error) on any
// storage failure.
func (s *Server) persistAgentNodeName(w http.ResponseWriter, r *http.Request, agentID, nodeName string) bool {
	if s.store == nil {
		return true
	}
	if err := s.store.UpdateAgentNodeName(r.Context(), agentID, nodeName); err != nil {
		s.logger.ErrorContext(r.Context(), "update agent node_name in store failed", "error", err)
		writeError(w, http.StatusInternalServerError, msgStorageError)
		return false
	}
	return true
}

// applyAgentRename mutates the in-memory agent record under the write
// lock. Returns the updated record + previous name, or ok=false (after
// writing the 404) when the agent disappeared between checks.
func (s *Server) applyAgentRename(w http.ResponseWriter, agentID, nodeName string) (Agent, string, bool) {
	s.mu.Lock()
	var oldName string
	agent, exists := s.updateAgentIdentity(agentID, func(a *Agent) {
		oldName = a.NodeName
		a.NodeName = nodeName
	})
	s.mu.Unlock()
	if !exists {
		writeError(w, http.StatusNotFound, msgAgentNotFound)
		return Agent{}, "", false
	}
	return agent, oldName, true
}

// agentDeregisterScope checks the URL/auth/scope preconditions for
// deregistering an agent and returns the chi-extracted agent id when allowed.
// On any failure it has already written the HTTP error response.
func (s *Server) agentDeregisterScope(w http.ResponseWriter, r *http.Request, user auth.User) (string, bool) {
	agentID := chi.URLParam(r, "id")
	if agentID == "" {
		writeError(w, http.StatusBadRequest, msgMissingAgentID)
		return "", false
	}
	preCheck, preExists := s.live.Get(agentID)
	if !preExists {
		writeError(w, http.StatusNotFound, msgAgentNotFound)
		return "", false
	}
	scope, ok := s.requireFleetScope(w, r, user)
	if !ok {
		return "", false
	}
	if !scope.IsAllowed(preCheck.FleetGroupID) {
		writeError(w, http.StatusNotFound, msgAgentNotFound)
		return "", false
	}
	return agentID, true
}

// persistAgentDeregister wipes the agent's persistent state. Returns false
// after writing an HTTP error if a fatal storage operation failed; the
// caller must abort. Recovery-grant revocation and revocation persistence
// are best-effort.
func (s *Server) persistAgentDeregister(w http.ResponseWriter, r *http.Request, agentID string, agent Agent) bool {
	if s.store == nil {
		return true
	}
	if _, err := s.store.RevokeAgentCertificateRecoveryGrant(r.Context(), agentID, s.now()); err != nil && !errors.Is(err, storage.ErrNotFound) {
		s.logger.ErrorContext(r.Context(), "revoke cert recovery grant failed", "agent_id", agentID, "error", err)
	}
	if err := s.store.DeleteInstancesByAgent(r.Context(), agentID); err != nil {
		s.logger.ErrorContext(r.Context(), "delete instances by agent failed", "agent_id", agentID, "error", err)
		writeError(w, http.StatusInternalServerError, msgStorageError)
		return false
	}
	// R9b: cancel any buffered agent upsert still sitting in the batch writer
	// so it can't flush a PutAgent AFTER the DeleteAgent below and resurrect
	// the row (the in-memory revocation guard only stops NEW snapshots).
	if s.batchWriter != nil {
		s.batchWriter.RemovePendingAgent(agentID)
	}
	if err := s.store.DeleteAgent(r.Context(), agentID); err != nil && !errors.Is(err, storage.ErrNotFound) {
		s.logger.ErrorContext(r.Context(), "delete agent from store failed", "agent_id", agentID, "error", err)
		writeError(w, http.StatusInternalServerError, msgStorageError)
		return false
	}
	// P1-SEC-06: persist the revocation so the ID stays rejected
	// across restarts until the underlying cert expires.
	certExpires := s.now().AddDate(0, 0, 30) // fallback to default lifetime if unknown
	if agent.CertExpiresAt != nil {
		certExpires = *agent.CertExpiresAt
	}
	if err := s.store.PutAgentRevocation(r.Context(), storage.AgentRevocationRecord{
		AgentID:       agentID,
		RevokedAt:     s.now().UTC(),
		CertExpiresAt: certExpires.UTC(),
	}); err != nil {
		s.logger.ErrorContext(r.Context(), "persist agent revocation failed", "agent_id", agentID, "error", err)
		// Non-fatal: in-memory revocation below still blocks the
		// current process. Restart recovery will see this as a gap.
	}
	// Drop any outstanding self-update desired-state: the agent is gone, so
	// nothing will ever satisfy it. The map is persisted, so a leaked key
	// would outlive the agent row forever.
	if s.updatesSvc != nil {
		if err := s.updatesSvc.ClearPendingAgentUpdate(r.Context(), agentID); err != nil {
			s.logger.WarnContext(r.Context(), "clear pending agent self-update on deregister failed",
				"agent_id", agentID, "error", err)
		}
		// Same reasoning for a pending telemt.update (Task 11) — independent
		// persisted key, same leaked-tombstone risk.
		if err := s.updatesSvc.ClearPendingTelemtUpdate(r.Context(), agentID); err != nil {
			s.logger.WarnContext(r.Context(), "clear pending telemt update on deregister failed",
				"agent_id", agentID, "error", err)
		}
	}
	return true
}

// markAgentRevoked arms the resurrection guard for agentID under s.mu. Called
// before the storage delete in handleDeregisterAgent so the guard is already
// active when RemovePendingAgent/DeleteAgent run and when a concurrent
// snapshot's enqueue path re-checks it. Idempotent — purgeAgentInMemory sets
// the same key again as part of the full in-memory teardown.
func (s *Server) markAgentRevoked(agentID string) {
	s.mu.Lock()
	s.revokedAgentIDs[agentID] = struct{}{}
	s.mu.Unlock()
}

// purgeAgentInMemory clears every in-memory map associated with the agent.
// Lock ordering: mu -> Service.mu (client usage lives in clients.Service).
func (s *Server) purgeAgentInMemory(agentID string) {
	s.mu.Lock()
	// live.Remove evicts the agent and all of its instances from the mirror.
	// Taken under s.mu (s.mu -> live) alongside the server-side map deletes.
	s.live.Remove(agentID)
	delete(s.detailBoosts, agentID)
	delete(s.initializationWatchCooldowns, agentID)
	// The self-update re-enqueue throttle is keyed by agent ID: without this a
	// re-registered agent inherits the dead one's throttle window.
	delete(s.selfUpdateReenqueuedAt, agentID)
	// Same reasoning for the telemt.update re-enqueue throttle (Task 11).
	delete(s.telemtUpdateReenqueuedAt, agentID)
	// Drop the agent's usage rows from the clients.Service mirror — the
	// single owner of usage state. Service.mu is acquired while holding
	// Server.mu, matching the documented Server.mu -> Service.mu lock ordering.
	if s.clientsSvc != nil {
		s.clientsSvc.DropAgentUsageMirror(agentID)
	}
	// Drop the last-known Telemt reachability so a re-registered agent starts
	// from the "first observation is never an edge" state instead of inheriting
	// a stale unreachable flag (P5, audit #18: Forget was dead because its
	// wiring was missing). ReachabilityTracker has its own lock and never calls
	// back into the server, so calling it under s.mu preserves lock ordering.
	s.telemtReach.Forget(agentID)
	// Drop the agent's runtime-events ring (P6-6.3e): without this the
	// per-agent ring leaks after deregistration and a reused agentID
	// would resurrect stale events.
	if s.runtimeEvents != nil {
		s.runtimeEvents.Remove(agentID)
	}
	s.revokedAgentIDs[agentID] = struct{}{}
	s.mu.Unlock()
}

func (s *Server) handleDeregisterAgent() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		session, user, err := s.requireSession(r)
		if err != nil {
			writeErrorLogged(r.Context(), w, http.StatusUnauthorized, "unauthorized", err)
			return
		}
		agentID, ok := s.agentDeregisterScope(w, r, user)
		if !ok {
			return
		}

		// 1. Signal the gRPC stream to shut down by closing the done channel.
		//    This is safe against concurrent notifyAgentSession because the
		//    notify path checks done before sending to wake. The session
		//    manager (controlplane/agents.SessionManager) encapsulates the
		//    map + close(done) bookkeeping; see P3-ARCH-01a.
		s.sessions.Terminate(agentID)

		// 2. Verify agent exists before doing any work.
		agent, exists := s.live.Get(agentID)
		if !exists {
			writeError(w, http.StatusNotFound, msgAgentNotFound)
			return
		}

		// 3. Arm the resurrection guard BEFORE the storage delete. Ordered this
		//    way so persistAgentDeregister's RemovePendingAgent/DeleteAgent run
		//    with the guard already active: a concurrent heartbeat is either
		//    dropped at the guard check in applyAgentSnapshot, or — if it already
		//    slipped past that check — has its buffered PutAgent re-checked
		//    against the guard under s.mu in enqueueAgentSnapshotBatchWrites, so
		//    it can never flush a PutAgent after the DeleteAgent below (R9b).
		s.markAgentRevoked(agentID)

		// 4. Persist deletion to storage so a failure does not leave the agent
		//    absent from memory but present in the database.
		if !s.persistAgentDeregister(w, r, agentID, agent) {
			return
		}

		// 5. Clean up the remaining in-memory state (mirror, per-agent rings,
		//    usage). The revocation flag was already set in step 3; setting it
		//    again here is idempotent.
		s.purgeAgentInMemory(agentID)

		// 6. Remove from presence tracker.
		s.presence.Remove(agentID)

		s.appendAuditWithContext(r.Context(), session.UserID, "agents.deregister", agentID, map[string]any{})

		w.WriteHeader(http.StatusNoContent)
	}
}

// handleUpdateAgentFleetGroup reassigns an agent to a different fleet
// group. Mirrors handleRenameAgent: scope-checks BOTH the current group
// (caller must already manage it) AND the target group (caller must be
// allowed to add the agent to it). Persists via the storage layer,
// patches the in-memory snapshot, and audit-logs the move with old/new
// group ids so reassignments are reversible from the audit trail.
func (s *Server) handleUpdateAgentFleetGroup() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		session, user, err := s.requireSession(r)
		if err != nil {
			writeErrorLogged(r.Context(), w, http.StatusUnauthorized, "unauthorized", err)
			return
		}

		agentID := chi.URLParam(r, "id")
		if agentID == "" {
			writeError(w, http.StatusBadRequest, msgMissingAgentID)
			return
		}

		var req updateAgentFleetGroupRequest
		if err := decodeJSON(r, &req); err != nil {
			writeErrorLogged(r.Context(), w, http.StatusBadRequest, "invalid request body", err)
			return
		}
		newGroupID := strings.TrimSpace(req.FleetGroupID)
		if newGroupID == "" {
			writeError(w, http.StatusBadRequest, "fleet_group_id is required")
			return
		}

		// Verify the agent exists and the caller is scoped over its
		// CURRENT group before any work — otherwise an out-of-scope
		// caller would learn the agent's existence via the 4xx code.
		existing, exists := s.live.Get(agentID)
		if !exists {
			writeError(w, http.StatusNotFound, msgAgentNotFound)
			return
		}

		scope, ok := s.requireFleetScope(w, r, user)
		if !ok {
			return
		}
		if !scope.IsAllowed(existing.FleetGroupID) {
			writeError(w, http.StatusNotFound, msgAgentNotFound)
			return
		}
		// Operator must also be allowed to add agents to the TARGET
		// group; otherwise a partial-scope user could exfiltrate an
		// agent into a group they otherwise cannot manage.
		if !scope.IsAllowed(newGroupID) {
			writeError(w, http.StatusForbidden, "fleet group out of scope")
			return
		}

		if existing.FleetGroupID == newGroupID {
			// No-op: return the current shape so the client can
			// refresh without a special-cased "no change" branch.
			writeJSON(w, http.StatusOK, existing)
			return
		}

		// Reject targeting a non-existent group up front rather than
		// surfacing the storage-layer FK error to the operator.
		if s.store != nil {
			if _, err := s.store.GetFleetGroup(r.Context(), newGroupID); err != nil {
				if errors.Is(err, storage.ErrNotFound) {
					writeErrorLogged(r.Context(), w, http.StatusNotFound, "fleet group not found", err)
					return
				}
				s.logger.ErrorContext(r.Context(), "get fleet group for reassign failed", "error", err)
				writeError(w, http.StatusInternalServerError, msgStorageError)
				return
			}
			if err := s.store.UpdateAgentFleetGroup(r.Context(), agentID, newGroupID); err != nil {
				if errors.Is(err, storage.ErrNotFound) {
					writeErrorLogged(r.Context(), w, http.StatusNotFound, msgAgentNotFound, err)
					return
				}
				s.logger.ErrorContext(r.Context(), "update agent fleet group in store failed", "error", err)
				writeError(w, http.StatusInternalServerError, msgStorageError)
				return
			}
		}

		// Patch the in-memory snapshot so subsequent /agents and
		// /fleet-groups reads reflect the move without waiting for the
		// next heartbeat. Captures oldGroupID for the audit event.
		s.mu.Lock()
		var oldGroupID string
		current, stillExists := s.updateAgentIdentity(agentID, func(a *Agent) {
			oldGroupID = a.FleetGroupID
			a.FleetGroupID = newGroupID
		})
		s.mu.Unlock()
		if !stillExists {
			writeError(w, http.StatusNotFound, msgAgentNotFound)
			return
		}

		s.appendAuditWithContext(r.Context(), session.UserID, "agents.reassign_fleet_group", agentID, map[string]any{
			"old_fleet_group_id": oldGroupID,
			"new_fleet_group_id": newGroupID,
		})

		// R10 (C4): the move changed which nodes each client targets. Roll the
		// clients of the group it joined onto this node, and pull the clients
		// of the group it left off it. Without this the panel claimed a
		// deployment the node had never heard of — and, worse, left the old
		// group's clients live on a node that is no longer supposed to serve
		// them.
		s.clientsSvc.ReconcileTopology(r.Context(), session.UserID, s.now())

		// P3: the new group may carry config the node has never applied
		// (its desired snapshot still reflects the old group, so the usual
		// applyDiff path would not surface the new group's sections). Push
		// them now so the node catches up on join instead of waiting for an
		// unrelated config-apply run. Best-effort: the reassign itself has
		// already succeeded and must not be rolled back over an apply
		// failure.
		if batchID, err := s.applyGroupToAgent(r.Context(), session.UserID, agentID, newGroupID); err != nil {
			s.logger.WarnContext(r.Context(), "auto-apply group config on reassign failed",
				"agent_id", agentID, "group_id", newGroupID, "error", err)
		} else if batchID != "" {
			s.appendAuditWithContext(r.Context(), session.UserID, "config.group_auto_apply", agentID,
				map[string]any{"group_id": newGroupID, "batch_id": batchID})
		}

		writeJSON(w, http.StatusOK, current)
	}
}

// Fleet-group HTTP handlers moved to http_fleet_groups.go as part of
// the groups redesign (UUID ids, CRUD + integrations). The previous
// handleFleetGroups derived the list from the live agent snapshot;
// the replacement reads the fleet_groups table so empty groups also
// appear in the list.