# openapi/panvex.yaml
...
openapi: 3.0.3
info:
  title: Panvex Control Plane API
  description: |
    Operator-facing HTTP API for the Panvex fleet management
    control-plane. Migrated incrementally from hand-rolled Zod
    schemas; consult `web/src/shared/api/schemas/` for endpoints not
    yet covered here.
  version: "1.0.0-pre"
  contact:
    name: Panvex maintainers
    url: https://github.com/lost-coder/panvex
servers:
  - url: /
    description: Same-origin (the panel serves both API and UI)
paths:
...
  /api/agents/{id}/fleet-group:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      operationId: updateAgentFleetGroup
      summary: Reassign an agent to a different fleet group
      description: |
        Operator must be scoped over BOTH the current and target
        groups. No-op responses (existing == target) are returned
        with the current agent shape so the client can refresh
        without a special-cased branch.
      tags: [agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentFleetGroupRequest'
      responses:
        '200':
          description: Agent reassigned (or no-op).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/agents/{id}/certificate-recovery-grants/revoke:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: revokeAgentCertificateRecoveryGrant
      summary: Revoke a previously-issued recovery grant
      description: |
        Marks the agent's recovery grant as revoked so a stale grant
        cannot be redeemed. Returns the updated grant snapshot.
      tags: [agents, recovery]
      responses:
        '200':
          description: Grant revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentCertificateRecoveryGrant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Error:
      type: object
      description: Standard error envelope used by every 4xx / 5xx response.
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code (e.g. `not_found`, `forbidden`).
        message:
          type: string
          description: Human-readable message; safe to display to operators.
        details:
          type: object
          description: Optional structured context. Shape varies per error code.
          additionalProperties: true
    Agent:
      type: object
      description: |
        Control-plane snapshot of one enrolled agent. Mirrors the Go
        `server.Agent` struct (`internal/controlplane/server/types.go`).
      required:
        - id
        - node_name
        - fleet_group_id
        - version
        - read_only
        - presence_state
        - runtime
        - last_seen_at
      properties:
        id:
          type: string
          description: Agent UUID.
        node_name:
          type: string
          description: Operator-facing display name.
        fleet_group_id:
          type: string
          description: UUID of the fleet group the agent belongs to.
        version:
          type: string
          description: Agent binary semver as last reported.
        read_only:
          type: boolean
          description: True when the agent serves read-only Telemt instances.
        presence_state:
          type: string
          description: |
            Live presence evaluation — `online`, `flapping`, `offline`, etc.
            Computed at request time from session + heartbeat data.
        transport_reconnect_pending:
          type: boolean
          description: |
            True when the agent was switched to outbound transport but the
            panel has not accepted a session from/to it since the switch
            ("switched but never reconnected"). Cleared on the next accepted
            agent stream.
        transport_drift:
          type: boolean
          description: |
            True when the direction of the last accepted stream disagreed with
            dial_transport_mode — the agent is still dialing IN while the DB
            says outbound, or vice versa (R-4). The panel re-enqueues the
            switch job to converge; cleared once a session connects in the
            direction the DB expects.
        bootstrap_state:
          type: string
          description: |
            Enrollment lifecycle state of the agents row: "pending"
            (provisioned, awaiting first enrollment), "expired" (bootstrap
            token lapsed before enrollment) or "active" (enrolled). Empty for
            legacy inbound rows. Meaningful for the UI only while the node has
            never connected — a half-added node renders as pending/expired
            instead of offline (R9a).
        dial_transport_mode:
          type: string
          description: |
            Persisted transport mode of the agents row — "inbound" (agent
            dials the panel) or "outbound" (panel dials the agent). Distinct
            from runtime.transport_mode (Telemt classic/middle_proxy).
        certificate_recovery:
          allOf:
            - $ref: '#/components/schemas/AgentCertificateRecoveryGrant'
          description: |
            Optional snapshot of any active certificate-recovery grant
            for the agent. Absent when no grant has been issued.
        cert_issued_at:
          type: string
          format: date-time
          description: Issuance timestamp of the agent's current certificate.
        cert_expires_at:
          type: string
          format: date-time
          description: Expiry timestamp of the agent's current certificate.
        telemt_update_probe:
          allOf:
            - $ref: '#/components/schemas/TelemtUpdateProbe'
          description: |
            Whether (and how) this node's telemt install can be updated
            in place: which process supervisor fronts it, and — when no
            in-place path exists — why. Cached by the agent once at
            process startup and stamped on every snapshot; absent until
            the agent has sent at least one such snapshot (older agents
            that predate the probe never populate it).
        runtime:
          $ref: '#/components/schemas/AgentRuntime'
        last_seen_at:
          type: string
          format: date-time
          description: Last successful heartbeat timestamp.
...
    AgentRuntime:
      type: object
      description: |
        Telemt operator overview reported by the agent. The set of
        fields is defensive on the Zod side because backends can ship
        new counters without an immediate web release; on the spec
        side we list every field the panel currently consumes.
      required:
        - accepting_new_connections
        - me_runtime_ready
        - me2dc_fallback_enabled
        - use_middle_proxy
        - startup_status
        - startup_stage
        - startup_progress_pct
        - initialization_status
        - degraded
        - initialization_stage
        - initialization_progress_pct
        - transport_mode
        - current_connections
        - current_connections_me
        - current_connections_direct
        - active_users
        - uptime_seconds
        - connections_total
        - connections_bad_total
        - handshake_timeouts_total
        - configured_users
        - dc_coverage_pct
        - healthy_upstreams
        - total_upstreams
        - fail_rate_pct_5m
        - fail_rate_known
        - connect_attempt_total
        - connect_success_total
        - connect_fail_total
        - connect_failfast_total
        - dcs
        - upstreams
        - recent_events
        - system_load
        - telemt_unreachable
        - telemt_unreachable_since_unix
        - user_telemetry_suppressed
        - updated_at
      properties:
        accepting_new_connections: { type: boolean }
        me_runtime_ready: { type: boolean }
        me2dc_fallback_enabled: { type: boolean }
        use_middle_proxy: { type: boolean }
        startup_status: { type: string }
        startup_stage: { type: string }
        startup_progress_pct: { type: number, format: double }
        initialization_status: { type: string }
        degraded: { type: boolean }
        lifecycle_state: { type: string }
        initialization_stage: { type: string }
        initialization_progress_pct: { type: number, format: double }
        transport_mode: { type: string }
        current_connections: { type: integer }
        current_connections_me: { type: integer }
        current_connections_direct: { type: integer }
        active_users: { type: integer }
        uptime_seconds: { type: number, format: double }
        connections_total: { type: integer, format: int64, minimum: 0 }
        connections_bad_total: { type: integer, format: int64, minimum: 0 }
        connections_bad_by_class:
          type: array
          nullable: true
          description: |
            Per-class breakdown of bad connections (Telemt 3.4.10+). The
            class set is open-ended. Backend emits `null` (not `[]`) when
            no breakdown has been observed yet — absence means "unknown".
          items:
            $ref: '#/components/schemas/ConnectionClassCount'
        handshake_failures_by_class:
          type: array
          nullable: true
          description: |
            Per-class breakdown of handshake failures (Telemt 3.4.10+).
            Same null-vs-empty semantics as connections_bad_by_class.
          items:
            $ref: '#/components/schemas/ConnectionClassCount'
        handshake_timeouts_total: { type: integer, format: int64, minimum: 0 }
        configured_users: { type: integer }
        dc_coverage_pct: { type: number, format: double }
        healthy_upstreams: { type: integer }
        total_upstreams: { type: integer }
        # Per-route-kind upstream counts (Telemt 3.4.7+, surfaced through
        # the gateway). Optional: older agents / Telemt versions don't
        # report them, so the panel falls back to counting the upstream
        # rows. The Go struct emits them without omitempty, but leaving
        # them out of `required` keeps the panel backward-compatible.
        unhealthy_upstreams: { type: integer, minimum: 0 }
        direct_upstreams: { type: integer, minimum: 0 }
        socks4_upstreams: { type: integer, minimum: 0 }
        socks5_upstreams: { type: integer, minimum: 0 }
        shadowsocks_upstreams: { type: integer, minimum: 0 }
        fail_rate_pct_5m:
          type: number
          format: double
          description: |
            5-minute upstream connect fail-rate. Read together with
            `fail_rate_known`: `false` means "unknown", not "0%".
        fail_rate_known: { type: boolean }
        connect_attempt_total: { type: integer, format: int64, minimum: 0 }
        connect_success_total: { type: integer, format: int64, minimum: 0 }
        connect_fail_total: { type: integer, format: int64, minimum: 0 }
        connect_failfast_total: { type: integer, format: int64, minimum: 0 }
        fallback_entered_at_unix:
          type: integer
          format: int64
          nullable: true
          description: |
            Unix timestamp the panel observed this agent enter
            ME→DC fallback. Absent when not in fallback.
        reroute_active: { type: boolean }
        route_mode: { type: string }
        me2dc_fast_enabled: { type: boolean }
        stale_cache_used: { type: boolean }
        top_by_connections:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeTopByConnections'
        top_by_throughput:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeTopByThroughput'
        dcs:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeDC'
        upstreams:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeUpstream'
        recent_events:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeEvent'
        system_load:
          $ref: '#/components/schemas/RuntimeSystemLoad'
        me_writers_summary:
          $ref: '#/components/schemas/RuntimeMeWritersSummary'
        telemt_unreachable: { type: boolean }
        telemt_unreachable_since_unix: { type: integer, format: int64 }
        # True when Telemt is not emitting per-user series (user telemetry
        # switched off, or the 4096-user labeled-series cap truncated them):
        # per-client traffic and quota usage for this node are stale or zero.
        # Warning only — does not affect node status. False is healthy.
        user_telemetry_suppressed: { type: boolean }
        updated_at:
          type: string
          format: date-time
...
    ConnectionClassCount:
      type: object
      description: |
        One (class, total) pair from Telemt's classified bad-connection
        and handshake-failure counters. Mirrors the Go
        `ConnectionClassCount` presentation type.
      required: [class, total]
      properties:
        class: { type: string }
        total: { type: integer, format: int64, minimum: 0 }
    RuntimeTopByConnections:
      type: object
      required: [username, connections]
      properties:
        username: { type: string }
        connections: { type: integer }
    RuntimeTopByThroughput:
      type: object
      required: [username, throughput_bytes]
      properties:
        username: { type: string }
        throughput_bytes: { type: integer, format: int64, minimum: 0 }
    AgentCertificateRecoveryGrant:
      type: object
      description: Lifecycle snapshot of a single recovery grant.
      required:
        - agent_id
        - status
        - issued_at_unix
        - expires_at_unix
      properties:
        agent_id: { type: string }
        status:
          type: string
          enum: [allowed, expired, used, revoked]
        issued_at_unix: { type: integer, format: int64 }
        expires_at_unix: { type: integer, format: int64 }
        used_at_unix:
          type: integer
          format: int64
          nullable: true
        revoked_at_unix:
          type: integer
          format: int64
          nullable: true
    TelemtUpdateProbe:
      type: object
      description: |
        Result of the agent probing its local process supervisor
        (systemd, OpenRC/procd, runit, docker, or none) once at startup,
        used by the panel to decide whether an in-place telemt update is
        safe to offer for this node.
      required:
        - mode
        - suggested_restart_spec
        - binary_path
        - available
        - reason
      properties:
        mode:
          type: string
          description: |
            "binary" (a supported supervisor owns telemt; a binary swap +
            supervised restart is safe), "docker" (telemt runs in a
            container; updates must go through the image), or "none"
            (nothing detected).
        suggested_restart_spec:
          type: string
          description: |
            telemtrestart spec ("systemd:telemt", "openrc:telemt",
            "procd:telemt", "runit:telemt") to use for the supervised
            restart after a binary swap. Empty when mode != "binary".
        binary_path:
          type: string
          description: |
            Best-effort resolved path to the telemt executable, used as
            the swap target. May be empty if it could not be resolved.
        available:
          type: boolean
          description: Whether an in-place update is offered at all.
        reason:
          type: string
          description: |
            Localizable CODE (not a human-readable phrase) explaining why
            available is false, e.g. "docker_only",
            "no_service_manager_detected". Empty when available is true.
...
    UpdateAgentFleetGroupRequest:
      type: object
      required: [fleet_group_id]
      properties:
        fleet_group_id:
          type: string
          minLength: 1
          description: Target fleet-group UUID. Must already exist.
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Caller is not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller is authenticated but lacks the required role/scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource does not exist or is out-of-scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: panvex_session
      description: |
        Server-issued session cookie set by `POST /api/login`. Marked
        Secure + HttpOnly + SameSite=Lax in production; the bare
        non-prefixed name above is the dev fallback for plain-HTTP
        loopback (the production runtime sets `__Host-panvex_session`
        instead — same auth flow, different cookie attribute set).
security:
  - sessionCookie: []